Apply suggestion from @atalman

+++ torchci/pages/api/oot/results.ts
@@ -34,7 +34,7 @@ export default async function handler(
       return res.status(401).json({ error: "Unauthorized" });
     }
     const a = Uint8Array(Buffer.from(raw));
-    const b = Buffer.from(expected);
+    const b = Uint8Array(Buffer.from(expected));
     if (a.length !== b.length || !timingSafeEqual(a, b)) {
       return res.status(401).json({ error: "Unauthorized" });
     }
